Permit predeclaring multiple size class queues
Right now it's only possible to predeclare the size class queue
belonging to the maximum size class of a platform. For certain setups it
may be desirable to autoscale all workers for all sizes down to zero. To
make that work, you want to predeclare multiple size classes.

This change replaces the maximumSizeClass field with a repeated field
that's simply named sizeClasses. If you don't care about multiple size
classes, you can simply specify the following in config:

    sizeClasses: [0]

Require that size classes are specified in ascending order, so that we
can easily ensure the provided values are unique.