Add an NFSv4.1 server
You see that most other NFS server implementations combine NFSv4.0 and
NFSv4.1 into a single server. For our use case this isn't necessary,
because we always know which exact minor version clients are going to
use. We can thus provide separate implementations.

Furthermore, I'm not convinced that combining both versions into a
single server is desirable from a code maintenance perspective. Even
though 4.1 is only 2.5% more than 4.0, the differences between both
protocols are quite substantial:

- The NFSv4.0 way of registering clients with SETCLIENTID and
  SETCLIENTID_CONFIRM has been completely replaced by EXCHANGE_ID and
  CREATE_SESSION.

- Transactions using sequence IDs on open-owners and lock-owners have
  been completely replaced by SEQUENCE.

- With NFSv4.0, the client ID has to be derived from state IDs, meaning
  state IDs are globally unique. With NFSv4.1, the client ID has to be
  derived from the sequence ID that is provided to SEQUENCE. State IDs
  are now scoped to a client.

With that in mind, this change adds a new server type that implements
NFSv4.1 with complete disregard for compatibility with NFSv4.0, though
it does borrow a lot of code from the NFSv4.0 server.

The protocol changes described above require a completely different
approach to locking. The NFSv4.0 server is protected by a single lock
that we only needed to acquire for a subset of requests. With NFSv4.1 we
need to perform some locking for every request, because almost all of
them will start with SEQUENCE. Fortunately, with state IDs now being
scoped to clients, it's easy to add fine-grained locking:

- Clients, sessions and slot are protected by mutex
  nfs41Program.clientsLock.

- Open-owners and lock-owners are protected by read-write lock
  clientIncarnationState.lock. Except when a client is fully idle, as
  then we can use nfs41Program.clientsLock.

- Opened files are protected by read-write lock
  nfs41Program.openedFilesLock.

- Byte-range lock sets are protected by read-write lock
  nfs41OpenedFileState.locksLock.