Set ExecuteOperationMetadata.digest_function
This field just got added to the protocol in the following PR:

https://github.com/bazelbuild/remote-apis/pull/311

For us it's trivial to expose this information, as we already have it
stored in the Remote Worker desired state message. There it is
guaranteed to be set, even for the traditional digest functions that
don't require it.