Replace GetContainingDigests() with VirtualApply()
Right now we have completely different code paths for computing the
containing digests for files and directories. For files we call
LinkableLeaf.VirtualApply(), while for directories we call
InitialContentsFetcher.GetContainingDigests().

The downside of the current approach is that generic virtual file system
APIs are strongly coupled against REv2. By giving InitialContentsFetcher
a VirtualApply(), we get rid of the coupling. In addition to that, it's
now possible to call InitialChild.GetNode().VirtualApply() to obtain the
set of containing digests in a file type oblivious way.