Replace FUSE's direct_mount option with mount_method
When enabling direct_mount, go-fuse attempts to create a FUSE mount
directly, using the mount(2) system call. If that fails, it falls back
to using the fusermount utility. What's annoying is that this causes any
errors returned by mount(2) to be suppressed. go-fuse solved this by
adding a DirectMountStrict option, which disables the fallback
behaviour.

This change exposes go-fuse's DirectMountStrict feature. Instead of
adding a separate boolean flag for it, we promote the existing
direct_mount configuration option to an enumeration named mount_method.
This means that if your configuration previously contained:

    directMount: true,

You need to replace it with either one of the lines below, depending on
whether you want to still fall back to calling fusermount:

    mountMethod: 'DIRECT',
    mountMethod: 'DIRECT_AND_FUSERMOUNT',

Fixes: #156