Let inMemoryPrepopulatedDirectory forward VirtualApply() calls
InitialContentsFetcher already has a VirtualApply() method, which is
used by bb_clientd to efficiently extract the set of containing digests
of a directory. This functionality is currently only available through
FilterChildren().

This change makes it possible to call into the InitialContentsFetcher
directly, without going through FilterChildren. This will be useful as
part of bonanza_worker, where we want to call into VirtualApply() to
prevent unnecessary computation of directory Merkle trees.