Enable AppleXcodeResolvingRunner on macOS unconditionally
Now that it supports setting XCODE_VERSION_OVERRIDE to DEVELOPER_DIR
paths, you may want to use this resolution logic even if no
apple_xcode_developer_directories are set in config. Because this code
has no effect if XCODE_VERSION_OVERRIDE is absent or DEVELOPER_DIR is
present, this should be fairly safe.