Disable support for REv2.0 output_{files,directories} by default
Command.output_paths was added to REv2 in November 2019. At the same
time output_files and output_directories were deprecated. At this point
there is no excuse for clients to use the old options. Let's place
support for the old fields behind a configuration option, so that we can
deorbit support for it early next year.