Workaround MSVC header inclusion issues
This worksaround this bug in bazel:
https://github.com/bazelbuild/bazel/issues/19733. In short, Bazel's
c++ header inclusion checking is broken when executing MSVC under remote
execution. This commit worksaround this issue for bzlmod-based builds
following a suggestion in that thread by making all builds run inside a
specific named subfolder. We thus change the input root to
execroot/_main on Windows, instead of the previous default of root.