Add support for NFSv4 named attributes (#195)
* Allow nodes to report the value of FATTR4_NAMED_ATTR

Up until this point we always let this attribute be false. However, if
we're going to implement OPENATTR then we should make sure that this
actually reflects whether one or more named attributes are present.

* Add Node.VirtualOpenNamedAttributes() for NFSv4 OPENATTR

This change alters the OPENATTR implementations in our NFSv4 servers to
call into a new method on the nodes, which can be used to look up the
named attributes directory. The next step will be to actually provide an
implementation of such a directory type.

* Give named attributes and their directories the right file type

I'd rather not extend filesystem.FileType to include these types, as
they are highly specific to NFSv4. Let's instead of that reuse
FileTypeDirectory and FileTypeRegularFile, while storing in a separate
flag whether the node is part of a named attribute directory.

* Let macOS write extended attributes in NFSv4 named attributes

* No document that AppleDouble files need to be filtered out

Now that we used NFSv4 named attributes, macOS will no longer attempt to
create AppleDouble files. This means that the filter can be simplified.

If we ever drop support for NFSv4.0, we should also remove support for
this configuration option entirely.

* Properly convert StatusErrWrongType to NFS4ERR_WRONG_TYPE

Right now we miss an entry for that, meaning that any accidental
attempts to call OPENATTR against a named attribute will not return the
right error code.

* Let VirtualBuildDirectory take DefaultAttributesSetter directly

Instead of passing in individual fields that we want to set in the
attributes, we might as well just pass in the function directly.

* Give mutable files and directories named attribute directories

This change replaces the stubs we added for VirtualOpenNamedAttributes()
to PoolBackedFileAllocator and InMemoryPrepopulatedDirectory with a
functional implementation. To facilitate this we add a new
NamedAttributes type that the former can embed to give a node named
attributes.

- For the worker's top-level directory we add NoNamedAttributesFactory.
  This one always returns EACCES when OPENATTR is called.

- For build directories we provide an implementation that creates
  instances of InMemoryPrepopulatedDirectory. To reduce memory usage we
  let all of those directories be backed by the same subtree object.

- To make sure that named attributes can't have named attributes
  themselves, we add the somewhat wacky named
  InNamedAttributeDirectoryNamedAttributesFactory. I lack the
  inspiration to come up with a better name. This one also ensures that
  nodes belonging to named attribute directories are of type
  NF4NAMEDATTR and NF4ATTRDIR instead of NF4REG and NF4DIR.