NFSv4: Let SETATTR respect the provided state ID (#196)
Right now these operations always behave as if the anonymous or read
bypass stateID is provided. This is probably good enough, but it has the
disadvantage that lease renewing is never performed. It also means we
don't return NFS4ERR_OLD_STATEID when necessary. In theory it may also
lead to cases where NFS4ERR_STALE is returned, which isn't correct for
files that are opened.

This change decomposes getOpenedLeaf(), so that we can reuse the logic
for looking up a leaf by state ID and validating the share access. We
don't want to reuse all of getOpenedLeaf(), as our VFS API does not
require opening a file to change its size.