pkg/filesystem: Make the zeroes underlying a pool file changeable
This allows bonanza to have (mutable) files backed by others. Bonanza
can use another FileReader to supply the data for all the "holes" in a
pool file.

Note that this requires some care to work correctly: If a file is
truncated to shrink and later extended again, the HoleSource must track
that and emit zeroes from offset of the smallest truncation.