pkg/filesystem: Make the zeroes underlying a pool file changeable (#200)
* pkg/filesystem: Rename variable pool to filePool (noop)

So a future commit can still access the package by its name

* pkg/filesystem: Make the zeroes underlying a pool file changeable

This allows bonanza to have (mutable) files backed by others. Bonanza
can use another FileReader to supply the data for all the "holes" in a
pool file.

Note that this requires some care to work correctly: If a file is
truncated to shrink and later extended again, the HoleSource must track
that and emit zeroes from offset of the smallest truncation.

* Make FilePoolStatsBuildingExecutor aware of non-zero initial file size

Right now we leave the size of the file set to zero, even if it was
created with a non-zero size. This causes the stats for these files to
be incorrect.

* Rename DefaultHoleSource to ZeroHoleSource

This is also the name of the underlying type.

* Document HoleSource

* Correct the implementation of GetNextRegionOffset()

Instead of simply returning the max() of the hole offsets of the file
and the hole source, we need to repeatedly check for holes in both
places. Only when we reach an offset where both locations correspond to
a hole, we may return.

* Ensure HoleSource.Truncate() is called properly

* Let WriteAt() read data from the HoleSource when writing partial sectors

Instead of writing null bytes, the write path must now also merge data
coming from the HoleSource. Otherwise we unintentially overwrite data
before/after the area to be written with null bytes.

* Don't use ZeroHoleSource as part of tests

I'd rather use mocking here, so that we ensure that we properly respect
the HoleSource instead of filling in null bytes ourselves.

---------

Co-authored-by: Ed Schouten <eschouten@apple.com>