Add JoinedSuspendable
I'm interested in making it possible for bb_worker to talk to separate
helper processes to determine whether the clock used for enforcing
execution timeouts needs to be suspended. Ideally we'd configure this
per worker thread. However, the processes I want to talk to won't be
able to provide information at that granularity.