Add support for execution timeout compensation via HTTP (#201)
* Factor out Suspendable

* Add JoinedSuspendable

I'm interested in making it possible for bb_worker to talk to separate
helper processes to determine whether the clock used for enforcing
execution timeouts needs to be suspended. Ideally we'd configure this
per worker thread. However, the processes I want to talk to won't be
able to provide information at that granularity.

* Add a HTTP client for suspending/resuming a Suspendable

* Add support for execution timeout compensation via HTTP

Right now bb_worker is capable of compensating the execution timeout for
any overhead caused by downloading files stored in the input root.
However, there are also certain workloads where actions need to access
files that are stored outside of the input root.

This change adds the ability for bb_worker to call into an external
process via HTTP to determine whether any overhead is taking place. The
protocol for this is simple: in bb_worker's configuration file you
specify two URLs. These are called in an alternating manner. The HTTP
server should let these calls block until the requested criterion is
reached (i.e., suspend and resume).