Let bb_scheduler announce correct API versions
In our case everything below REv2.3 is deprecated. The implementation is
expected to be compatible with any client built on top of the latest
released version of protocol, which is REv2.11.

While there, push back the date at which we're going to remove support
for the actionAndCommand platform key extractor. That way we can remove
all of these deprecated features in one go.