Remove support for REv2.0 output files/directories
On 2019-11-19, REv2 changed the way expected output paths are declared
in the Command message. It used to be the case that files and
directories were declared separately, but this prevents you from writing
an action that captures either of those, depending on what the action
generated.

Given that this feature was added 6.5 years ago, let's go ahead and
remove support for the old scheme. Clients have had more than enough
time to implement the new scheme.