Remove support for platform properties in Command messages
REv2.2 promoted the platform properties from the Command message to the
Action message. This has the advantage that the scheduler no longer
needs to process any Command messages. This saves one round trip to the
CAS. Command messages can in some cases also become very large, meaning
it adds unnecessary memory pressure on the scheduler.

Given that this feature was added on 2020-10-17 and clients were able to
adopt quickly. it's time to remove support for the old scheme.