Forbid multiple size classes with the lowest one being zero
For size class queues that are created dynamically, we already enforce
that if a platform queue uses size classes it is no longer possible to
create size class queue 0.

However, for ones that are predeclared we are missing such a check. This
means that it's possible for people to create a platform queue having
size class queues 0 and 1. This is meaningless, as a worker can't be
zero times as large as another one.

Reported by: Daniel Thagard via Slack